+++ rog-control-center/src/pages/mod.rs
@@ -3,9 +3,3 @@ mod app_settings;
 mod aura_page;
 mod fan_curve_page;
 mod system_page;
-
-pub use anime_page::*;
-pub use app_settings::*;
-pub use aura_page::*;
-pub use fan_curve_page::*;
-pub use system_page::*;