+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Fix to suspend process in anime thread to let custom anims run on wake
 
 ## [v5.0.6]
 - Revert egui update due to a lot of issues arising from window closing.